Move test keypad context wrapper into `testing/` as that's the only ones that use it now
