+++ testing/test-keypad-context-wrapper.tsx
@@ -1,12 +1,13 @@
+import {View} from "@khanacademy/wonder-blocks-core";
+import {action} from "@storybook/addon-actions";
+import {StyleSheet} from "aphrodite";
+import * as React from "react";
+
 import {
     KeypadContext,
     StatefulKeypadContextProvider,
 } from "@khanacademy/keypad-context";
 import {MobileKeypad} from "@khanacademy/math-input";
-import {View} from "@khanacademy/wonder-blocks-core";
-import {action} from "@storybook/addon-actions";
-import {StyleSheet} from "aphrodite";
-import * as React from "react";
 
 const Footer = (): React.ReactElement => {
     return (
